chexpert.py- removed resize, fixed crop to 320, removed H flip removed resize to 256 from qwp if restored restore option
my_models.py- removed double from forward pass
Psi_PhiTMP.py - added relative BSU file path
wpt_transform.py -
add vis_channels as self parameter
in __call__ added dimension check to remove added 3rd dim
resize to 256*256 is now inside __call__
in transform_image - removed np.array

+++ models/my_models.py
@@ -90,7 +90,7 @@ class WPTResNet(models.ResNet):
 
         # x = self.maxpool(x) # without maxpool, then wptdownsample is needed in pre-processing
 
-        x = self.layer1(x).double()
+        x = self.layer1(x)
         x = self.layer2(x)
         x = self.layer3(x)
         x = self.layer4(x)
